+++ js/main.js
@@ -1,5 +1,7 @@
 import { createOffers } from './create-offers.js';
 import './popup.js';
-import './offer-form.js';
+import './validation/validation-rooms-guests.js';
+import './validation/validation-price-title.js';
+
 
 createOffers();
