+++ js/validation/validation-price-title.js
@@ -1,8 +1,7 @@
 const offerForm = document.querySelector('.ad-form');
 const offerTitle = offerForm.querySelector('#title');
 const offerPrice = offerForm.querySelector('#price');
-const offerRoomNumber = offerForm.querySelector('#room_number');
-const offerCapacity = offerForm.querySelector('#capacity');
+
 
 const pristine = new Pristine(offerForm, {
   classTo: 'ad-form__element',
@@ -13,17 +12,6 @@ const pristine = new Pristine(offerForm, {
   errorTextClass: 'form__error'
 });
 
-const selectRooms = offerForm.querySelector('[name="rooms"]');
-const selectCapacity = offerForm.querySelector('[name="capacity"]');
-const roomsOptions = {
-  '1': ['1'],
-  '2': ['1', '2'],
-  '3': ['1', '2', '3'],
-  '100': ['0'],
-};
-
-const validateRoomsOptions = () => roomsOptions[selectRooms.value].includes(selectCapacity.value);
-
 const titleLength = (value) => value.length >= 30 && value.length <= 100;
 const priceMaxValue = (number) => number <= 100000;
 const priceMinValue = (number) => number > 0;
@@ -31,10 +19,11 @@ const priceMinValue = (number) => number > 0;
 pristine.addValidator(offerTitle, titleLength, 'От 30 до 100 символов');
 pristine.addValidator(offerPrice, priceMaxValue, 'Максимальная стоимость не может превышать 100 000');
 pristine.addValidator(offerPrice, priceMinValue, 'Минимальная стоимость должна быть больше 0');
-pristine.addValidator(offerRoomNumber, validateRoomsOptions, 'Количество гостей не должно превышать количество комнат');
-pristine.addValidator(offerCapacity, validateRoomsOptions, 'Количество гостей не должно превышать количество комнат');
+
 
 offerForm.addEventListener('submit', (evt) => {
   evt.preventDefault();
   pristine.validate();
 });
+
+export {offerForm, pristine};
